refactor(agent): extract SWM-sender-state serialisation from dkg-agent.ts
Moves the six SWM sender-key key/serialisation helpers
(`swmSenderStateKey`, `swmReceiverStateKey`,
`(de)serializeSwmSenderSendState`, `(de)serializeSwmSenderReceiveState`)
out of the bottom of `dkg-agent.ts` into a focused
`dkg-agent-swm-state.ts` module, plus the three module-private
validators (`requiredString`, `optionalString`, `requiredNumber`) that
back the deserialisers.

These functions are pure transformations between in-memory
`LocalSwmSenderKey{Send,Receive}State` and the on-disk JSON shape —
no `DKGAgent` state, no class dependency.

Final tally for the dkg-agent.ts Phase 0 reduction in this PR:

  packages/agent/src/dkg-agent.ts         14,453 → 13,485 lines (-968)
  packages/agent/src/dkg-agent-constants.ts            +155 lines
  packages/agent/src/dkg-agent-types.ts                +578 lines
  packages/agent/src/dkg-agent-helpers.ts              +389 lines
  packages/agent/src/dkg-agent-swm-state.ts            +128 lines

All extractions are 1:1 mechanical moves; the public surface
(`@origintrail-official/dkg-agent` package barrel) is unchanged and
`dkg-agent.ts` re-exports the moved public symbols for any caller that
still imports directly from `./dkg-agent.js`.
